resize avatar picker images in worker threads
for each call to Layout, create a thumbnail cache and populate it from
the last thumbnail map for each of the keys (filenames) that belong in
the thumbnail cache by iterating over the filenames. If an operation is
already busy, wait for it to complete before scheduling more thumbnails
to the workers. When the cache grows too large, prune the entries and
let gc free memory of thumbnail images.