start replacing catshadow
This compiles and connects but doesn't do much else yet.

The idea is to replace the home view with a list of type Conversation,
which holds messages from Contacts who tag their Message with the Conversation id.

Each Contact will have a Stream via which Messages are sent.

To be implemented: saving state to disk, panda or reunion integration