update gio API usage
system.StageEvent is deprecated and system.FrameEvent is moved to app.FrameEvent.

This means we need to keep track of whether we're getting FrameEvents to learn whether the application has focus on Android, in order to start a Foreground service.
There might be other event types that we can monitor to learn this. TBD.